refactor BudgetService - 更新了 BudgetService 接口和实现类中的方法签名，将 Integer 类型的参数改为 Long 类型 - 新增了 getCurrentMonthBudgetUsage 和 getMonthBudgetUsage 方法 -优化了预算提醒和预算状态检查的逻辑 - 删除了未使用的 CategoryKeywordMapping 相关代码- 移除了未使用的 category-keyword-mapping.sql 文件

+++ src/main/java/com/acco/life/controller/AccountTransactionController.java
@@ -50,7 +50,7 @@ public class AccountTransactionController {
 
     @Operation(summary = "根据 ID 查询 AccountTransaction")
     @GetMapping("/{id}")
-    public Mono<ResponseEntity<AccountTransactionDto>> get(@PathVariable Integer id) {
+    public Mono<ResponseEntity<AccountTransactionDto>> get(@PathVariable Long  id) {
         return service.findById(id).map(ResponseEntity::ok)
         .onErrorResume(e ->
         Mono.just(ResponseEntity.badRequest().build()));
@@ -74,7 +74,7 @@ public class AccountTransactionController {
 
     @Operation(summary = "删除 AccountTransaction")
     @DeleteMapping("/{id}")
-    public Mono<Void> delete(@PathVariable Integer id) {
+    public Mono<Void> delete(@PathVariable Long  id) {
         return service.deleteById(id);
     }
 }
