+++ src/main/java/com/acco/life/controller/AssetAccountController.java
@@ -51,7 +51,7 @@ public class AssetAccountController {
 
     @Operation(summary = "根据 ID 查询 AssetAccount")
     @GetMapping("/{id}")
-    public Mono<ResponseEntity<AssetAccountDto>> get(@PathVariable Integer id) {
+    public Mono<ResponseEntity<AssetAccountDto>> get(@PathVariable Long  id) {
         return service.findById(id).map(ResponseEntity::ok)
         .onErrorResume(e ->
         Mono.just(ResponseEntity.badRequest().build()));
@@ -79,7 +79,7 @@ public class AssetAccountController {
 
     @Operation(summary = "删除 AssetAccount")
     @DeleteMapping("/{id}")
-    public Mono<Void> delete(@PathVariable Integer id) {
+    public Mono<Void> delete(@PathVariable Long  id) {
         return service.deleteById(id);
     }
