+++ src/main/java/com/acco/life/controller/InvestmentAssetController.java
@@ -49,7 +49,7 @@ public class InvestmentAssetController {
 
     @Operation(summary = "根据 ID 查询 InvestmentAsset")
     @GetMapping("/{id}")
-    public Mono<ResponseEntity<InvestmentAssetDto>> get(@PathVariable Integer id) {
+    public Mono<ResponseEntity<InvestmentAssetDto>> get(@PathVariable Long id) {
         return service.findById(id).map(ResponseEntity::ok)
         .onErrorResume(e ->
         Mono.just(ResponseEntity.badRequest().build()));
@@ -73,7 +73,7 @@ public class InvestmentAssetController {
 
     @Operation(summary = "删除 InvestmentAsset")
     @DeleteMapping("/{id}")
-    public Mono<Void> delete(@PathVariable Integer id) {
+    public Mono<Void> delete(@PathVariable Long id) {
         return service.deleteById(id);
     }
 }
